pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "cython", "numpy>=1.19.3", "pysam>=0.21.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gwplot"
version = "0.3.0"
license = { file = "LICENSE" }
description = "Visualise genomics data fast"
requires-python = ">=3.9"
readme = "README.md"
authors = [
    { name = "Kez Cleal", email = "clealk@cardiff.ac.uk" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]
dependencies = [ "setuptools>=61.0" ]

[project.optional-dependencies]
test = ["matplotlib",
    "Flask",
    "pysam>=0.21.0",
    "numpy>=1.19.3",
    "pillow"]
pysam = ["pysam>=0.21.0"]
numpy = ["numpy>=1.19.3"]
pillow = ["pillow"]
all = ["pysam>=0.21.0",
    "numpy>=1.19.3",
    "pillow"]

[tool.setuptools]
packages = ["gwplot"]

[project.urls]
Repository = "https://github.com/kcleal/gwplot"
